apps/qa_worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = ""
authors = ["王博伟 <wangbowei.114514@bytedance.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
farm-haystack = { extras = ["faiss"], version = "^1.15.1" }
numpy = "1.23.0"
pika = "^1.3.1"
pydantic = "^1.10.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
